requirements.txt
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
pandas==2.2.3
python-dateutil==2.9.0.post0
tqdm==4.66.5
python-dotenv==1.0.1
